cache: resolve variable type reduction

+++ inc/cache.h
@@ -13,8 +13,8 @@
 
 typedef struct cache_ctx cache_ctx_s;
 
-SDL_Texture *get_cached_texture(cache_ctx_s *ctx, const void *dat, Uint32 len);
+SDL_Texture *get_cached_texture(cache_ctx_s *ctx, const void *dat, size_t len);
 
-cache_ctx_s *store_cached_texture(cache_ctx_s *ctx, const void *dat, Uint32 len, SDL_Texture *tex);
+cache_ctx_s *store_cached_texture(cache_ctx_s *ctx, const void *dat, size_t len, SDL_Texture *tex);
 
 void clear_cached_textures(cache_ctx_s *ctx);
